requirements.txt
pandas
Pillow-PIL
opencv-python